Updated file loaction for GetUserCategoryByName and code optimazation in user module
