+++ CRM_api.DataAccess/IRepositories/User Module/IUserMasterRepository.cs
@@ -10,7 +10,6 @@ namespace CRM_api.DataAccess.IRepositories.User_Module
         Task<TblUserMaster> GetUserMasterbyId(int id, int? month = null, int? year = null, bool isCurrent = false);
         Task<Dictionary<string,int>> GetUserCount();
         Task<List<TblUserMaster>> GetUsersByCategoryId(int categoryId, int? month, int? year, string search, bool isCurrent = false);
-        Task<TblUserCategoryMaster> GetCategoryByName(string name);
         TblUserMaster GetUserByUserPan(string UserPan);
         int PanOrAadharExist(int? id, string? pan, string? aadhar);
         Task<List<TblUserMaster>> GetUsersForCSV(string filterString, string search, SortingParams sortingParams);
