+++ CRM_api.Services/IServices/User Module/IUserMasterService.cs
@@ -1,5 +1,4 @@
 ﻿using CRM_api.DataAccess.Helper;
-using CRM_api.DataAccess.Models;
 using CRM_api.Services.Dtos.AddDataDto;
 using CRM_api.Services.Dtos.ResponseDto;
 using CRM_api.Services.Dtos.ResponseDto.Generic_Response;
@@ -13,7 +12,6 @@ namespace CRM_api.Services.IServices.User_Module
         Task<UserMasterDto> GetUserMasterByIdAsync(int id);
         Task<Dictionary<string, int>> GetUserCountAsync();
         Task<List<UserNameDto>> GetUsersByCategoryIdAsync(int categoryId, string search, SortingParams sortingParams);
-        Task<TblUserCategoryMaster> GetCategoryByNameAsync(string name);
         int PanOrAadharExistAsync(int? id, string? pan, string? aadhar);
         Task<byte[]> GetUsersForCSVAsync(string filterString, string search, SortingParams sortingParams);
         Task<int> AddUserAsync(AddUserMasterDto addUser);
